Work only in portrait up

+++ lib/resources/values_manager.dart
@@ -27,8 +27,8 @@ class AppSize {
   static const double s0_25 = 0.25;
   static const double s1 = 1.0;
   static const double s1_5 = 1.5;
-  static const int s2 = 2;
-  static const int s3 = 3;
+  static const double s2 = 2;
+  static const double s3 = 3;
   static const double s4 = 4.0;
   static const double s8 = 8.0;
   static const double s10 = 10.0;
